Modernize codebase and fix compiler fallout
Squashes 23 original commits from 89f41c8d..8a4e67c0.
First: Add AGENTS.md and PacketPool, avoid range copies
Last:  Add missing include for std::array in StarHostAddress

Representative commits:
- 89f41c8d Add AGENTS.md and PacketPool, avoid range copies
- b5c90187 Avoid range copies in game subsystems
- 1bf57bf8 Optimize world collision cache and entity iteration
- e49e6740 Fix narrowing conversions and pass drawable lists by const ref
- e82a7167 CI: Upgrade to GCC 14
- 8a4e67c0 Add missing include for std::array in StarHostAddress

+++ source/frontend/StarCharCreation.cpp
@@ -247,7 +247,7 @@ void CharCreationPane::changed() {
   auto speciesDefinition = Root::singleton().speciesDatabase()->species(m_speciesList[m_speciesChoice]);
   auto species = speciesDefinition->options();
   auto genderOptions = species.genderOptions.wrap(m_genderChoice);
-  int genderIdx = pmod<int64_t>(m_genderChoice, species.genderOptions.size());
+  int genderIdx = pmod<int64_t>(m_genderChoice, static_cast<int64_t>(species.genderOptions.size()));
 
   auto labels = speciesDefinition->charGenTextLabels();
